+++ src/jwt.rs
@@ -8,6 +8,7 @@ use crate::{config::ConfJWT, models::common::UserID};
 
 /// 30 days
 const LIFETIME_SEC: u64 = 30 * 24 * 60 * 60;
+const JWT_AUD_PREFIX: &str = "/resource";
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Claims {
@@ -21,7 +22,7 @@ pub fn generate(user_id: UserID, config: &ConfJWT) -> Result<String> {
     Ok(encode(
         &Header::default(),
         &Claims {
-            aud: config.audience.clone(),
+            aud: config.issuer.clone() + JWT_AUD_PREFIX,
             iss: config.issuer.clone(),
             user_id,
             exp: get_current_timestamp() + LIFETIME_SEC,
@@ -33,7 +34,7 @@ pub fn generate(user_id: UserID, config: &ConfJWT) -> Result<String> {
 pub fn validate(token: &str) -> Result<UserID> {
     let config = crate::get_config()?;
     let mut validation = Validation::default();
-    validation.set_audience(&[&config.jwt.audience]);
+    validation.set_audience(&[config.jwt.issuer.clone() + JWT_AUD_PREFIX]);
     validation.set_issuer(&[&config.jwt.issuer]);
 
     let token = decode::<Claims>(